inline DisplayPanel into HUD
